make "edit" button functional
In it's current state, a user can search a song, load and read it,
and after that open up the song in editor. Yet, the editor is still
incomplete. Line-based approach is bad, so I'll replace it with a
single text box (DIV).

Apart from the editor, only missing main feature is adding new
songs. But a complete rewrite/re-organisation of code may also
be appropriate.